# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: SML Staff Marketplace API
  version: "1.1.0"
  description: |
    Express.js port of the original Java/JAX-RS MarketPlaceWebService.

    All endpoints are under `/service/v1` (included in the server URL below).

    **Auth note**: Failed login returns HTTP 200 with `data: []` — not HTTP 401.

    **Image endpoints**: Return binary `image/png`. `/imagesguid` never returns 404 — returns empty buffer instead.
servers:
  - url: http://localhost:47309/service/v1
    description: Local development (current SML Staff Ubon)
  - url: http://localhost:47300/service/v1
    description: Local development (legacy preview)
...
security:
  - GUIDHeader: []
components:
  securitySchemes:
    GUIDHeader:
      type: apiKey
      in: header
      name: GUID
      description: Session identifier required by most endpoints
  schemas:
    ErrorResponse:
      type: object
      properties:
        ERROR:
          type: string
          example: "database connection error"
...
    CustomerCreditData:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        credit_day:
          type: integer
          example: 30
        credit_date:
          type: string
          format: date
          description: Computed as NOW() + credit_day days
    AdvancePayment:
      type: object
      properties:
        doc_no:
          type: string
        doc_date:
          type: string
          format: date
        total_amount:
          type: number
        used:
          type: number
        balance_amount:
          type: number
...
    ImageMeta:
      type: object
      properties:
        image_id:
          type: string
          description: Same as item_code
        guid_code:
          type: string
          format: uuid
...
paths:
...
  /getCustomerCredit:
    get:
      tags: [Financial]
      summary: ข้อมูลเครดิตของลูกค้า
      description: credit_date = NOW() + credit_day วัน (คำนวณใน DB)
      parameters:
        - { name: cust_code, in: query, required: true, schema: { type: string }, example: "C00001" }
      responses:
        '200':
          description: ข้อมูลเครดิต หรือ `{}` ถ้าไม่พบลูกค้า
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    oneOf:
                      - $ref: '#/components/schemas/CustomerCreditData'
                      - type: object
                        description: Empty object when not found
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /getAdvancePayment:
    get:
      tags: [Financial]
      summary: ประวัติมัดจำ/เงินล่วงหน้า
      description: |
        trans_flag IN (40, 9040) — max 20 รายการล่าสุด

        ถ้า last_status=1 → ยอดทั้งหมดแสดงเป็น 0
      parameters:
        - { name: cust_code, in: query, required: true, schema: { type: string }, example: "C00001" }
      responses:
        '200':
          description: รายการมัดจำ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/AdvancePayment' }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /getTotalBalance:
    get:
      tags: [Financial]
      summary: ยอดหนี้คงค้างรวม
      description: Invoice (44) + DN (46,93,99,95,101) - CN (48,97,103) - การชำระ ณ date(NOW())
      parameters:
        - { name: cust_code, in: query, required: true, schema: { type: string }, example: "C00001" }
      responses:
        '200':
          description: ยอดหนี้รวม
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:       { type: boolean }
                  total_balance: { type: number, example: 15840.50 }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ─── DOCUMENT ────────────────────────────────────────────────────────────────
...
  /getImageList:
    get:
      tags: [Images]
      summary: รายการ image metadata ของสินค้า
      description: ดึงจาก demo_images database — เรียงตาม image_order ASC
      security: []
      parameters:
        - { name: item_code, in: query, required: true, schema: { type: string }, example: "ITM001" }
      responses:
        '200':
          description: รายการ guid ของรูปภาพ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ImageMeta' }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /images:
    get:
      tags: [Images]
      summary: ดาวน์โหลดภาพตาม item_code
      description: |
        ส่งคืน binary PNG ของรูปแรกของสินค้า

        รองรับ ETag caching — ส่ง `If-None-Match` เพื่อรับ 304
        Cache-Control: `public, max-age=604800` (7 วัน)
      security: []
      parameters:
        - { name: item_code,      in: query,  required: true, schema: { type: string },         example: "ITM001" }
        - { name: If-None-Match,  in: header,                 schema: { type: string },         description: "ETag จาก request ก่อนหน้า" }
      responses:
        '200':
          description: PNG image binary
          headers:
            ETag:
              schema: { type: string }
              example: '"d41d8cd98f00b204e9800998ecf8427e"'
            Cache-Control:
              schema: { type: string }
              example: "public, max-age=604800"
            Content-Length:
              schema: { type: integer }
          content:
            image/png:
              schema: { type: string, format: binary }
        '304':
          description: Not Modified — client cache ยังใช้ได้
          headers:
            ETag:
              schema: { type: string }
            Cache-Control:
              schema: { type: string }
        '400':
          description: item_code ไม่ได้ส่งมา
          content:
            text/plain:
              schema: { type: string }
              example: "ERROR: item_code is required"
        '404':
          description: ไม่พบรูปภาพ
          content:
            text/plain:
              schema: { type: string }
              example: "ERROR: image not found"
  /imagesguid:
    get:
      tags: [Images]
      summary: ดาวน์โหลดภาพตาม guid_code
      description: |
        ใช้สำหรับรูปที่ 2, 3... ของสินค้า (ได้ guid จาก `/getImageList`)

        **ต่างจาก /images**: ถ้าไม่พบ guid → HTTP 200 + **empty buffer (0 bytes)** ไม่ใช่ 404

        ตรวจ `Content-Length: 0` ใน client เพื่อรู้ว่ารูปไม่มี
      security: []
      parameters:
        - { name: guid_code,      in: query,  required: true, schema: { type: string, format: uuid }, example: "550e8400-e29b-41d4-a716-446655440000" }
        - { name: If-None-Match,  in: header,                 schema: { type: string } }
      responses:
        '200':
          description: |
            PNG image binary — หรือ empty buffer (Content-Length=0) ถ้าไม่พบ
          headers:
            ETag:
              schema: { type: string }
              description: ไม่มีถ้า empty buffer
            Cache-Control:
              schema: { type: string }
            Content-Length:
              schema: { type: integer }
              description: 0 ถ้าไม่พบรูป
          content:
            image/png:
              schema: { type: string, format: binary }
        '304':
          description: Not Modified
          headers:
            ETag:
              schema: { type: string }
            Cache-Control:
              schema: { type: string }

  # ─── POS ─────────────────────────────────────────────────────────────────────
...